refactor: move storage stuff
