+++ artifacts/checksums.txt
@@ -1,4 +1,4 @@
 e2a0010c52831b3181cdee29a543adc5f608a8460a8065f68f71029d22b0bbe0  ibc_transfer.wasm
 d71f6f2a112af7ac4f4e711373ef161239cfb99b8b49348c3ef5d1bfaf219890  neutron_interchain_queries.wasm
-a3a9cb54a3b185bb013aacf647566af261106975b47355ad4b0b5d9bd9ee64f1  neutron_interchain_txs.wasm
+3b43d7283e36a36a888b444539f98c351d3cf19f795e57865635d9b1640eddea  neutron_interchain_txs.wasm
 cf66ca7c6f7027538671f1c0ea547f35f71dbfcc50907bc0bfb41a0782950dab  reflect.wasm
