+++ contracts/neutron_interchain_txs/src/storage.rs
@@ -1,4 +1,13 @@
+use cosmwasm_std::{from_binary, to_vec, Binary, StdResult, Storage};
 use cw_storage_plus::Map;
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub struct SudoPayload {
+    pub message: String,
+}
 
 pub const IBC_SUDO_ID_RANGE_START: u64 = 1_000_000_000;
 pub const IBC_SUDO_ID_RANGE_SIZE: u64 = 1_000_000;
@@ -8,3 +17,40 @@ pub const REPLY_QUEUE_ID: Map<u64, Vec<u8>> = Map::new("reply_queue_id");
 pub const SUDO_PAYLOAD: Map<(String, u64), Vec<u8>> = Map::new("sudo_payload");
 pub const INTERCHAIN_ACCOUNTS: Map<String, Option<(String, String)>> =
     Map::new("interchain_accounts");
+
+pub fn get_next_id(store: &mut dyn Storage) -> StdResult<u64> {
+    REPLY_QUEUE_ID
+        .keys(store, None, None, cosmwasm_std::Order::Descending)
+        .next()
+        .unwrap_or(Ok(IBC_SUDO_ID_RANGE_START))
+        .map(|id| id + 1)
+}
+
+pub fn save_reply_payload(store: &mut dyn Storage, payload: SudoPayload) -> StdResult<u64> {
+    let id = get_next_id(store)?;
+    REPLY_QUEUE_ID.save(store, id, &to_vec(&payload)?)?;
+    Ok(id)
+}
+
+pub fn read_reply_payload(store: &mut dyn Storage, id: u64) -> StdResult<SudoPayload> {
+    let data = REPLY_QUEUE_ID.load(store, id)?;
+    from_binary(&Binary(data))
+}
+
+pub fn read_sudo_payload(
+    store: &mut dyn Storage,
+    channel_id: String,
+    seq_id: u64,
+) -> StdResult<SudoPayload> {
+    let data = SUDO_PAYLOAD.load(store, (channel_id, seq_id))?;
+    from_binary(&Binary(data))
+}
+
+pub fn save_sudo_payload(
+    store: &mut dyn Storage,
+    channel_id: String,
+    seq_id: u64,
+    payload: SudoPayload,
+) -> StdResult<()> {
+    SUDO_PAYLOAD.save(store, (channel_id, seq_id), &to_vec(&payload)?)
+}
